use single 'isDevServerRunning' implementation
