+++ packages/plugin-metro/src/lib/start/OpenDebuggerKeyboardHandler.ts
@@ -3,10 +3,6 @@
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
- *
- * @flow strict-local
- * @format
- * @oncall react_native
  */
 
 import { color } from '@rnef/tools';
