+++ packages/tools/src/index.ts
@@ -14,6 +14,7 @@ export * from './lib/build-cache/remoteBuildCache.js';
 export * from './lib/build-cache/github/config.js';
 export * from './lib/build-cache/github/GitHubBuildCache.js';
 export { findDevServerPort } from './lib/dev-server/findDevServerPort.js';
+export { isDevServerRunning } from './lib/dev-server/isDevServerRunning.js';
 export { isInteractive } from './lib/isInteractive.js';
 export { spawn, SubprocessError } from './lib/spawn.js';
 export { color } from './lib/color.js';
