+++ packages/tools/src/lib/dev-server/findDevServerPort.ts
@@ -1,6 +1,6 @@
 import { handlePortUnavailable } from './handlePortUnavailable.js';
 import { isDevServerRunning } from './isDevServerRunning.js';
-import { logAlreadyRunningBundler } from './port.js';
+import { logAlreadyRunningBundler } from './logAlreadyRunningBundler.js';
 
 export const findDevServerPort = async (
   initialPort: number,
@@ -14,10 +14,7 @@ export const findDevServerPort = async (
 
   const devServerStatus = await isDevServerRunning(port);
 
-  if (
-    typeof devServerStatus === 'object' &&
-    devServerStatus.status === 'running'
-  ) {
+  if (devServerStatus.status === 'running') {
     if (devServerStatus.root === root) {
       startDevServer = false;
       logAlreadyRunningBundler(port);
@@ -25,7 +22,7 @@ export const findDevServerPort = async (
       const result = await handlePortUnavailable(port, root);
       [port, startDevServer] = [result.port, result.startDevServer];
     }
-  } else if (devServerStatus === 'unrecognized') {
+  } else if (devServerStatus.status === 'unrecognized') {
     const result = await handlePortUnavailable(port, root);
     [port, startDevServer] = [result.port, result.startDevServer];
   }
